# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Relationship Service",
        "title": "Relationship Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/relationships/friends/request": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send a friend request to another user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "relationships"
                ],
                "summary": "Send friend request",
                "parameters": [
                    {
                        "description": "Friend request payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_relationship_delivery_http_send_friend_request.SendFriendRequestHttp"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_relationship_delivery_http_send_friend_request.SendFriendRequestHttp": {
            "type": "object",
            "required": [
                "receiver_id"
            ],
            "properties": {
                "receiver_id": {
                    "type": "string"
                }
            }
        }
    }
}